chore(cli): Phase 3 — reorganize src/ into commands/, runtime/, internal-plugins/
Move src/bin/commands/* into src/commands/ (single-file commands at the
top level, multi-file commands like generate/migrate/drift/check get
their own folder with command.ts as the entry). Move the rest of
src/bin/* (utility modules) into src/runtime/, and src/bin/internal-plugins/
into src/internal-plugins/. src/bin/chkit.ts stays put as the actual
executable entry point.

Absorb src/drift.ts and src/drift-diff.ts (pure diff library) into
src/commands/drift/ as compare.ts and diff.ts so the drift logic lives
with its only consumer.

Extract buildDriftPayload from the drift command into a sibling
commands/drift/payload.ts so check.ts imports the shared domain
function without reaching into another command runner.

Mirror the new layout under src/test/ (test/runtime, test/commands,
test/internal-plugins). All imports updated; tsconfig and CLAUDE.md
paths refreshed accordingly. Pure restructuring — no behaviour change.
