+++ packages/cli/src/bin/chkit.ts
@@ -2,24 +2,24 @@
 import process from 'node:process'
 import type { ParsedFlags } from '@chkit/core'
 
-import { createCommandRegistry } from './command-registry.js'
-import { parseCommandArgs, runResolvedCommand } from './command-dispatch.js'
-import { generateCommand } from './commands/generate.js'
-import { migrateCommand } from './commands/migrate.js'
-import { statusCommand } from './commands/status.js'
-import { driftCommand } from './commands/drift.js'
-import { checkCommand } from './commands/check.js'
-import { pluginCommand } from './commands/plugin.js'
-import { queryCommand } from './commands/query.js'
-import { cmdInit } from './commands/init.js'
-import { loadConfig } from './config.js'
-import { GLOBAL_FLAGS } from './global-flags.js'
-import { formatGlobalHelp, formatCommandHelp } from './help.js'
-import { loadPluginRuntime } from './plugin-runtime.js'
-import { getInternalPlugins } from './internal-plugins/index.js'
-import { CLI_VERSION } from './version.js'
-import { debug } from './debug.js'
-import { configureCliLogging } from './logging.js'
+import { createCommandRegistry } from '../runtime/command-registry.js'
+import { parseCommandArgs, runResolvedCommand } from '../runtime/command-dispatch.js'
+import { generateCommand } from '../commands/generate/command.js'
+import { migrateCommand } from '../commands/migrate/command.js'
+import { statusCommand } from '../commands/status.js'
+import { driftCommand } from '../commands/drift/command.js'
+import { checkCommand } from '../commands/check/command.js'
+import { pluginCommand } from '../commands/plugin.js'
+import { queryCommand } from '../commands/query.js'
+import { cmdInit } from '../commands/init.js'
+import { loadConfig } from '../runtime/config.js'
+import { GLOBAL_FLAGS } from '../runtime/global-flags.js'
+import { formatGlobalHelp, formatCommandHelp } from '../runtime/help.js'
+import { loadPluginRuntime } from '../runtime/plugin-runtime.js'
+import { getInternalPlugins } from '../internal-plugins/index.js'
+import { CLI_VERSION } from '../runtime/version.js'
+import { debug } from '../runtime/debug.js'
+import { configureCliLogging } from '../runtime/logging.js'
 
 const WELL_KNOWN_PLUGIN_COMMANDS: Record<string, string> = {
   codegen: 'Codegen',
