+++ packages/cli/src/commands/generate/plan-pipeline.ts
