+++ packages/cli/src/commands/generate/rename-mappings.ts
