+++ packages/cli/src/internal-plugins/index.ts
@@ -1,4 +1,4 @@
-import type { ChxPlugin } from '../../plugins.js'
+import type { ChxPlugin } from '../plugins.js'
 
 import { createSkillHintPlugin } from './skill-hint.js'
