+++ packages/cli/src/runtime/command-registry.ts
