+++ packages/cli/src/runtime/table-scope.ts
