+++ packages/cli/src/test/runtime/safety-markers.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, test } from 'bun:test'
 
-import { extractExecutableStatements } from '../../bin/safety-markers.js'
+import { extractExecutableStatements } from '../../runtime/safety-markers.js'
 
 describe('extractExecutableStatements', () => {
   test('splits simple statement batches', () => {
